fix(jail): add /bin/sh symlink and register ax-jail bin target
Both jails now create /bin/sh → ~/bin/bash inside the bwrap sandbox via
--symlink, so Claude Code's hook runner (Node.js posix_spawn) can find
a shell on NixOS where /bin does not exist.

Also adds the missing [[bin]] ax-jail entry to Cargo.toml (was causing
nix build failures) and fixes three pre-existing clippy errors in the
ingest lib that blocked -D warnings.

+++ src/ingest/code.rs
@@ -387,7 +387,7 @@ fn make_chunks(source: &str, language: &str) -> Vec<Chunk> {
     }
 
     let total_lines = source.lines().count();
-    let max_covered = syms.iter().map(|s| s.end_line as usize).max().unwrap_or(0);
+    let max_covered = syms.iter().map(|s| s.end_line).max().unwrap_or(0);
 
     let mut chunks: Vec<Chunk> = syms
         .into_iter()
